Make Content interface and move logic to ContentRegular
